% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCarbon.R
\name{calcSortieC}
\alias{calcSortieC}
\title{Title}
\usage{
calcSortieC(outputs, dead = FALSE, BEC)
}
\arguments{
\item{outputs}{sortie output saved in a data.table}

\item{dead}{include snags in the calculation}

\item{BEC}{Biogeoclimatic ecosystem classification zone}
}
\description{
Title
}
\details{
Currently only supports common species in the SBS and ICH
To do: add argument for seedling to call the seedling carbon function
}
