% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCarbon.R
\name{cwdCfromSORTIE}
\alias{cwdCfromSORTIE}
\title{Coarse woody debris carbon}
\usage{
cwdCfromSORTIE(volGrid, spGroups)
}
\arguments{
\item{spGroups}{}
}
\description{
To do - make a default species group table and throw errors if species don't match
}
