% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSORTIE.R
\name{readSortie}
\alias{readSortie}
\title{Read SORTIE outputs}
\usage{
readSortie(filenames)
}
\arguments{
\item{filenames}{expecting a single csv for each site/plot with all timesteps}
}
\description{
Read SORTIE outputs
}
