% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCarbon.R
\name{calcSortieC}
\alias{calcSortieC}
\title{Title}
\usage{
calcSortieC(sortie_outputs, dead = FALSE, BEC)
}
\arguments{
\item{sortie_outputs}{sortie output saved in a data.table}

\item{dead}{include snags in the calculation}

\item{BEC}{Biogeoclimatic ecosystem classification zone}

\item{Ht_from_diam}{calculate the height based on the diameter (allometry) == TRUE
or use Height from Sortie == FALSE}
}
\description{
Title
}
\details{
Currently only supports common species in the SBS and ICH
To do: add argument for seedling to call the seedling carbon function
}
