% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSORTIE.R
\name{readSortie}
\alias{readSortie}
\title{read SORTIE}
\usage{
readSortie(outputs)
}
\arguments{
\item{outputs}{expecting a single csv for each site/plot with all timesteps or a vector of
filenames to read in}
}
\description{
read SORTIE
}
